pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "forge-studio"
version = "0.1.0"
description = "Web UI for the Forge inference control plane"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [{ name = "ArkaD171717" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "forge-infer[openai]",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "websockets>=14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.28",
    "ruff>=0.9",
]

[project.scripts]
forge-studio = "backend.cli:main"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
